fix mpv crash versions
