+++ app/src/main/java/eu/kanade/tachiyomi/util/CrashLogUtil.kt
@@ -5,7 +5,6 @@ import android.os.Build
 import eu.kanade.tachiyomi.BuildConfig
 import eu.kanade.tachiyomi.extension.anime.AnimeExtensionManager
 import eu.kanade.tachiyomi.extension.manga.MangaExtensionManager
-import eu.kanade.tachiyomi.ui.main.MainActivity
 import eu.kanade.tachiyomi.util.storage.getUriCompat
 import eu.kanade.tachiyomi.util.system.WebViewUtil
 import eu.kanade.tachiyomi.util.system.createFileInCacheDir
@@ -21,6 +20,8 @@ class CrashLogUtil(
     private val mangaExtensionManager: MangaExtensionManager = Injekt.get(),
     private val animeExtensionManager: AnimeExtensionManager = Injekt.get(),
 ) {
+
+    private val mpvVersions = MpvVersionsUtil(context).mpvVersions
 
     suspend fun dumpLogs() = withNonCancellableContext {
         try {
@@ -49,9 +50,9 @@ class CrashLogUtil(
             Device name: ${Build.DEVICE} (${Build.PRODUCT})
             Device model: ${Build.MODEL}
             WebView: ${WebViewUtil.getVersion(context)}
-            MPVLib version: ${MainActivity.mpvVersions.mpvCommit} (${MainActivity.mpvVersions.buildDate})
-            Libplacebo version: ${MainActivity.mpvVersions.libPlacebo}
-            FFmpeg version: ${MainActivity.mpvVersions.ffmpeg}
+            MPVLib version: ${mpvVersions.mpvCommit} (${mpvVersions.buildDate})
+            Libplacebo version: ${mpvVersions.libPlacebo}
+            FFmpeg version: ${mpvVersions.ffmpeg}
         """.trimIndent()
     }
